* Moved IDocument interface and related or extending interfaces from BaseTypes to DocumentTypes file. * Moved promoteDocumentsToCreatable(...) function from BaseTypes to DocumentTypes file. * Moved assertDocumentIsValid(...) function from Firebase file to DocumentTypes file.
